Refactor ONNX conversion function to make unit tests easier
Since the conversion script had a hardcoded directory in it, that made
the script fail when running it from the unit test class, which is in a
different directory. The conversion script is now a function that takes
in the directory path. This isn't as modular as before, but that's not a
huge deal; this is a one-off for now.