Make nnetInference() fixed-point compatible
This is likely just a first pass. I doubt that the lookup tables and
data types are optimized. If we run into problems fitting on the FPGA,
then we'll need to revisit this.

The labels predicted by the fixed-point version match the labels
predicted by the original model (see unit test in NnetTests).