[bug fix] Avoid divide by zero; indexing bug
Divide by zero:
When a harmonic isn't found, it's value is set to 0. This is problematic
when computing the ratios between harmonic fetures because it results in
a divide by 0. This has been fixed.

Indexing bug:
I always divided by harmonic*Ratio(1,n) instead of harmonic*Ratio(i,n).
This has been fixed...