use 1.5 power instead of square root cubed
There is a precision difference between the two, for some reason.
This precision difference was causing my feature extraction tests to
fail by a fairly large margin, sometimes up 10% difference.

However, I'm pretty sure 1.5 power won't be supported in HDL or
fixed-point, so I'll likely have to go back to the sqrt method and deal
with the precision difference.