HDL-compatible median function
The normal median function isn't supported because it uses sorting.

codegenMedian assumes that the input has an even number of elements,
which is true in our case (512). It is not very efficient, and will
probably really suck in hardware. It only supports vectors, not
matrices, so the median absolute deviation test was updated.