Trying to make HDL fft codegen work
The HDL FFT system object can't be called within a loop. The
documentation and a MATLAB Answers question indicate that the system
object needs to be in it's own function, then you can call that
function. Doing that is still not codegen compatible because I need to
generate the code that does all the validIn/validOut and buffering in a
loop; HDl Coder still complains about calling the system object in a for
loop.